@@ -35,16 +35,16 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   sqflite: ^2.0.0+3
-  intl: ^0.17.0
-  bloc: ^7.0.0
+  intl: ^0.18.0
+  bloc: ^8.1.1
   conditional_builder_null_safety: ^0.0.6
-  flutter_bloc: ^7.0.0
-  dio: ^4.0.0
-  hexcolor: ^2.0.5
+  flutter_bloc: ^8.1.2
+  dio: ^5.0.0
+  hexcolor: ^3.0.1
   shared_preferences: ^2.0.5
-  webview_flutter: ^2.0.3
+  webview_flutter: ^4.0.4
   smooth_page_indicator: ^1.0.0+2
-  toast: ^0.2.9
+  toast: ^0.3.0
   carousel_slider: ^4.1.1
 
 dev_dependencies:
