@@ -191,23 +191,18 @@ class ShopCubit extends Cubit<ShopStates>{
 
    }
 
-   bool isDark = false;
-   void changeMode({bool? fromShared})
-   {
-     if(fromShared != null)
-     {
+bool isDark=true;
+   void changeMode({bool? fromShared}) {
+     if (fromShared != null) {
        isDark = fromShared;
        emit(AppChangeCurrentModeState());
      }
-     else
-     {
+     else {
        isDark = !isDark;
        CacheHelper.saveData(key: 'isDark', value: isDark).then((value) {
          emit(AppChangeCurrentModeState());
        });
      }
-
    }
-
 
 }
