@@ -1,3 +1,6 @@
 import 'package:flutter/material.dart';
 
-const defaultColor = Color(0xff437dbb);
+const defaultColor = Colors.blue;
+
+/*
+Color(0xff437dbb)*/
